requirements/base.txt
fastapi==0.78.0
uvicorn==0.17.6
SQLAlchemy==1.4.36
sqlalchemy_utils==0.38.2
psycopg2-binary==2.9.3
databases==0.5.5
alembic==1.7.7
python-dotenv==0.20.0
cleo==0.8.1
email-validator==1.2.1
pyjwt==2.4.0
